add mod fchk: dump fch with librest2fch.so
